+++ bdtsim/protocol/simplepayment/simplepayment.py
@@ -21,7 +21,7 @@ from bdtsim.contract import SolidityContract
 from bdtsim.data_provider import DataProvider
 from bdtsim.environment import Environment
 from bdtsim.account import Account
-from bdtsim.protocol import Protocol, ProtocolManager
+from bdtsim.protocol import Protocol, ProtocolManager, DEFAULT_ASSET_PRICE
 from bdtsim.protocol_path import ProtocolPath
 
 
@@ -31,7 +31,20 @@ logger = logging.getLogger(__name__)
 class SimplePayment(Protocol):
     """Example implementation of SimplePayment (paying after exchange, direct payments)"""
     def execute(self, protocol_path: ProtocolPath, environment: Environment, data_provider: DataProvider,
-                seller: Account, buyer: Account, price: int = 1000000000) -> None:
+                seller: Account, buyer: Account, price: int = DEFAULT_ASSET_PRICE) -> None:
+        """
+
+        Args:
+            protocol_path (ProtocolPath): Protocol path this simulation for a data exchange will take
+            environment (Environment): The environment (blockchain) in which the protocol interactions will take place
+            data_provider (DataProvider): The data to be traded
+            seller (Account): The selling party
+            buyer (Account): The buying party
+            price (int): Price for the data/asset to be traded (in Wei)
+
+        Returns:
+            None
+        """
         release_goods = protocol_path.decide(seller, 'release goods?', ['yes', 'no'])
         # [...] seller releases goods to buyer, not monitored by BDTsim
         if release_goods == 'yes':
@@ -58,13 +71,39 @@ class AbstractParameterizedSimplePayment(Protocol):
             environment.deploy_contract(operator, self.get_contract())
 
     def execute(self, protocol_path: ProtocolPath, environment: Environment, data_provider: DataProvider,
-                seller: Account, buyer: Account, price: int = 1000000000) -> None:
+                seller: Account, buyer: Account, price: int = DEFAULT_ASSET_PRICE) -> None:
+        """
+
+        Args:
+            protocol_path (ProtocolPath): Protocol path this simulation for a data exchange will take
+            environment (Environment): The environment (blockchain) in which the protocol interactions will take place
+            data_provider (DataProvider): The data to be traded
+            seller (Account): The selling party
+            buyer (Account): The buying party
+            price (int): Price for the data/asset to be traded (in Wei)
+
+        Returns:
+            None
+        """
         raise NotImplementedError()
 
 
 class SimplePaymentPrepaid(AbstractParameterizedSimplePayment):
     def execute(self, protocol_path: ProtocolPath, environment: Environment, data_provider: DataProvider,
-                seller: Account, buyer: Account, price: int = 1000000000) -> None:
+                seller: Account, buyer: Account, price: int = DEFAULT_ASSET_PRICE) -> None:
+        """
+
+        Args:
+            protocol_path (ProtocolPath): Protocol path this simulation for a data exchange will take
+            environment (Environment): The environment (blockchain) in which the protocol interactions will take place
+            data_provider (DataProvider): The data to be traded
+            seller (Account): The selling party
+            buyer (Account): The buying party
+            price (int): Price for the data/asset to be traded (in Wei)
+
+        Returns:
+            None
+        """
         if protocol_path.decide(buyer, description='Payment', variants=['paying', 'not paying']).is_honest():
             logger.debug('Decided to be honest')
             if self._use_contract:
@@ -79,7 +118,20 @@ class SimplePaymentPrepaid(AbstractParameterizedSimplePayment):
 
 class SimplePaymentPostpaid(AbstractParameterizedSimplePayment):
     def execute(self, protocol_path: ProtocolPath, environment: Environment, data_provider: DataProvider,
-                seller: Account, buyer: Account, price: int = 1000000000) -> None:
+                seller: Account, buyer: Account, price: int = DEFAULT_ASSET_PRICE) -> None:
+        """
+
+        Args:
+            protocol_path (ProtocolPath): Protocol path this simulation for a data exchange will take
+            environment (Environment): The environment (blockchain) in which the protocol interactions will take place
+            data_provider (DataProvider): The data to be traded
+            seller (Account): The selling party
+            buyer (Account): The buying party
+            price (int): Price for the data/asset to be traded (in Wei)
+
+        Returns:
+            None
+        """
         if protocol_path.decide(seller, 'Give goods', ['yes', 'no']).is_honest():
             if protocol_path.decide(buyer, description='Payment', variants=['paying', 'not paying']).is_honest():
                 logger.debug('Decided to be honest')