+++ bdtsim/simulation.py
@@ -22,7 +22,7 @@ from typing import Optional
 from bdtsim.account import Account
 from bdtsim.data_provider import DataProvider
 from bdtsim.environment import Environment
-from bdtsim.protocol import Protocol
+from bdtsim.protocol import Protocol, DEFAULT_ASSET_PRICE
 from bdtsim.protocol_path import ProtocolPath, ProtocolPathCoercion
 from bdtsim.output import SimulationResult, ResultCollector
 
@@ -33,7 +33,7 @@ logger = logging.getLogger(__name__)
 class Simulation(object):
     def __init__(self, protocol: Protocol, environment: Environment, data_provider: DataProvider, operator: Account,
                  seller: Account, buyer: Account, protocol_path_coercion: Optional[ProtocolPathCoercion] = None,
-                 price: int = 1000000000) -> None:
+                 price: int = DEFAULT_ASSET_PRICE) -> None:
         self._protocol = protocol
         self._environment = environment
         self._data_provider = data_provider